move fetch_channel_id_values() from Dataset to DataProduct
